refactor/simplifications:
  - kill the ability to layer a netty channel on top of brokers.
nobody uses it.  everybody uses the Service interface.
  - change Broker into a Service[AnyRef, AnyRef] (we may move casting
up the stack soon)

this gets rid of write completion futures, which actually ends up
simplifying much of the semantics of finagle as we only have one
future per request, and there's a strict 1:1 relationship between
requests & futures.  in order to accomodate for this, we define
certain futures to be "write failures" so that we can match on them
for retries.

we also introduce improvements & fix some bugs.