Updated scaladoc to add package.scala info and updated README.md
Added new package.scala files at com and twitter to the point people to
the Finagle Intro and modified package.scala in finagle-core to contain
the intro. Updated package.scala in exception to point to exceptions
people can use. Moved placement of scaladoc in ServerBuilder.scala so
that it appears. Updated Finagle Developer Guide (README.md) to contain
a Java Patterns section, which includes a reference to external Java
example on Future thread pools.