[split] finagle-redis: api v2
This is phase I of changing finagle-redis' API. In this diff, I aimed to clean up interface inconsistencies, use ChannelBuffers, and minimize the amount of String/Array[Byte]/ChannelBuffer conversions.  In phase II, I will be replacing the codec with a re-written one, or Sam Pullara's codec, in order to get rid of Naggati completely and gain some more performance.

RB_ID=82347