[split] finagle-thrift: validate connections based on TApplicationException
Another real gem in Apache Thrift: it seems that servers can be left
in an undefined state if they emit certain TApplicationException — not
exactly what that exception class sounds like, is it? — WITHOUT
actually closing that connection. So, to account for egregious flaw,
we'll need to compensate in our client.

Also, this brings back test IDLs from f557572~1, since it's difficult
to make sense of the tests without them (I don't know why that change
removed them).

RB_ID=82884