[split] Added JSSE SSL support for intermediate certificates.
JSSE SSL previously only looked at a certificate and a key, not	an
intermediate certificate. Now the stack takes a caCertificate
argument. If there are multiple intermediate certificates,
cat them together as a single file.

One test spins up a server, connects with "openssl s_client" called
from the chain, and ensures that the (test) certificate chain is
valid.

It includes a tool that generates a new certificate chain for this
testing.

Another change: the JSSE cache of SSL instances used to be keyed
by the cert only. Now it's keyed by cert+key+chain (so now if we
bring up a second server with the same key but a different chain,
it will instantiate a second server rather than use the cached one,
which is IMHO correct).

Design notes:
- Previously JSSE SSL config took 2 args; it didn't use all 5 that the
  true OpenSSL config did. Now we're using a 3rd arg, the intermediate
  certificate, and are not passing through the remaining two args. The
  cert chain is used preferentially to the cert if the cert chain is
  specified.
- The test for "is the intermediate certificate used" was "null",
  which seemed weird to me, but I went with it.
- No awesome abstraction for calling a command from a shell and
  getting back the stdout/stderr streams.

RB_ID=87718