[split]  Add OstrichExporter to enable Ostrich stats in the new App stack.
For backward compatibility reasons, we add a http entry in the admin server
names "/stats" that return the stats computed by the ostrich library.
To enable this feature you just have to add a dependency to finagle-ostrich4,
we use the resources/META-INF/services mechanism to register this handler
into the admin server.
Note that, stats willl also be available in "/metrics.json" but computed
by the metrics library.

RB_ID=122706