[split]  Always enable TracingFilter even when no tracers are used.
An earlier commit to disable TracingFilter causes a problem when
an intermediate service has tracing disabled. spanId is embedded
into thrift messages so when the TracingFilter is disabled the
thrift reply message has the same spanId of the request message
which causes the sender to think it receives a message from itself.
Reverting the earlier commit before we find a better solution.

RB_ID=144828
TBR=true