[split]  Allow finagle-mysql to deal with non-ascii utf-8 characters
When pre-allocating the array for holding all the params for the PreparedStatement, finagle-mysql was using the assumption that each character is of 1 byte, which is not necessarily the case for UTF-8.  Need to use the number of bytes for the string to calculate the size for that array.

RB_ID=146514