[split] make thriftmux server back compatible with thrift framed codec
Added a ChannelPipelineFactory for thriftmux which can do the
translation between thrift and mux messages. The server switches
to backward-compatible mode if the first message is an invalid
mux message but a valid thrift message. This works reliably with
thrift framed codec + binary protocol + strict write. A thrift
message sent from the described client starts with a header
0x800100xx and mux does not confuse it with a valid mux message
since 0x80 is aninvalid mux message type. The server also works
with Finagle thrift clients. It answers the protocol
up-negotiation request and passed the tracing information embedded
in the subsquent thrift requests to mux which has native tracing
support.

RB_ID=147377