[split] Finagle client/server APIs: s/target/addr/g
Provide uniform naming that makes sense for both client and server.
This is also what Wily calls such names -- they are all forms of
addresses.

("Target" implies a "goal" -- address is more concrete. Not to mention
that address is more standard: most people could answer "what is an
address;" a "target" is not as well defined.)

RB_ID=165976