[split] finagle-mysql: better type support
The BINARY and VARBINARY types are similar to CHAR and VARCHAR,
except that they contain binary strings rather than nonbinary strings.
That is, they contain byte strings rather than character strings. This
caused a problem because finagle-mysql conflated the types. This patch
fixes that and improves handling of types by moving the decoding
of all non-primitive types to extractors.

RB_ID=223281
TBR=true