[split] Higher fidelity addrs from ZK We do the best we can with the current server set implementation. A new one is coming soon; this should be enough to limp along until then.
* This also sneaks in names for uses of the legacy
Resolver.resolve interfaces by creating a type of
Group which is little more than a container for a Name.

* Also caches resolutions down to env/role/job.

* Phrases stabilizing group in terms of Addrs.

Note: the goal with StabilizingAddr is to replicate
exactly what StabilizingGroup did. I don???t want to
change it. Something new and simpler is coming soon
enough. StabilizingAddr is transient.

RB_ID=219996