[split] finagle-core: adds a "tries" scoped statsfilter to ClientBuilder
Jeff Hodges pointed out that it's nontrivial to get stats on a given request to the services created by ClientBuilder.  The metrics that we get out of StatsFilter only give us information about each request we try to send over the wire.  Notably, he's interested in when the number of failures that the service actually ends up seeing (and not ones that are masked by finagle retries) start going up.

This is useful to service owners for two reasons.

1.  Service owners like alerts that are symptomatic and targeted.  If they are woken up in the middle of the night, it should be because something has a real problem, and they should know immediately what is on fire.

2.  When service owners configure ClientBuilders, they are doing it with a little wiggle room in mind.  They should be able to tolerate slightly higher downstream latency.  They should be able to tolerate a small bump in failed requests.  All of these are fine, up until a point, and that point is configured via retries, requestTimeout, etc.  They should be able to set up alerts so that they don't get alerts unless something is so bad that it's getting past the ClientBuilder wiggle room.

This change adds a StatsFilter that wraps the entire RetryingFilter if a retry policy has been set.  If users of finagle want information on these kinds of stats, they can now get them in the "tries" scope.

RB_ID=262013