[split] Option to enable thread pool scheduler in finagle, and fix the shutting down RejectedExecutionException's.
The option is in finagle exp now, users will be able to enable this without code change.

Don't have a fix for setUnsafe conflict when the user is also using setUnsafe. The one comes later will be used as the final scheduler.
And don't have a fix for errors caused by a executor with a bounded capacity.

RB_ID=266095