[split] finagle-mux: Adds lease support to mux clients Motivation: By getting out the client-side support for leases quickly, clients will be ready for leases when the servers start turning them on.
Luckily, the bulk of the work for leases is server-side, and Marius' client-side code was pretty clean.

The premise behind Acting is that Service is lossy at this layer--isAvailable really reflects whether the connection is open or not, and we want to know whether the connection is open or not, but also whether the lease is applied or not.  Depending upon the answer, we're either going to throw away the Service (to get a new connection) or we're just going to knock ourselves out of the load balancer for a little while.  isAvailable doesn't stop being lossy until the ServiceFactory layer, at which point all we care about is whether it is currently safe to send requests, and we'll stop sending requests if it's no longer available, so LeasedFactory connects the lossy Service isAvailable abstraction to the hermetic ServiceFactory.

Tlease will support a units field so that we can experiment with different varieties of leases.  Tlease messages should be infrequent, so it's fine to add a few extra bytes.

Modifications:
Adds Tlease to mux on the client side.
Adds an Acting trait, which when mixed into a Service, improves the fidelity to the point where we can describe being Acting.  This will probably be subsumed by Probation or TemporarilyUnavailable in the future.
Adds a LeasedBridger which uses the Service with Acting to form a proper ServiceFactory abstraction.
Changes ReusingPool to respect ServiceFactory availability
Adds tests

Results:
Mux clients will support Tlease in preparation for servers sending lease information.

RB_ID=277783