[split] finagle-core: Record Finagle version in tracing information
In order to be able to easily audit Finagle version in deployed services, we'd
like to record versions in Zipkin trace data at runtime. This change adds an
`AtomicReference[String]` to com.twitter.finagle.Init that is set with the
finagle-core version pulled from build.properties. This version is recorded
under the key "finagle.version" in com.twitter.finagle.tracing.TracingFilter.

RB_ID=284857