[split] Closing one side of a proxy should close the other
Motivations:

An HTTP proxy B between client A and server C.

    A<>B<>C

When A<>B closes we want B<>C to also close. Steps to reproduce are
described in the description of
https://jira.twitter.biz/browse/MACAW-316.

Modifications:

For a service definition:

> val client = ClientBuilder(..)
> val proxy = Service.mk { req => client(req) }

1. If we interrupt the client request, it should close the channel
between the client and a destination. To illustrate:

> client(req).raise(..)
> verify(ch, times(1)).close()

N.B. We can't quite do this because we don't have a reference to the
underlying Channel ch, and even if we did, we don't call close on it
directly but do it through Channels.close (why do we do this, is it to
avoid deadlocks?):

2. If we interrupt the proxy request, it should interrupt the client
request.

> val promise = new Promise[Response]
> val proxy = Service.mk { _ => promise }
> proxy(Request()).raise(..)
> assert(promise.isInterrupted.isDefined)

Now imagine a related service definition:

> val service = Service.mk { req =>
>   def loop = req.response.writer.write(..) before loop
>   loop
>   req.response.setChunked(true)
>   Future.value(req.response)
> }

3. Transport closure in the service dispatcher should discard the
reader.

> val req = Request()
> req.response.setChunked(true)
> val service = Service.mk { _ => Future.value(req.response) }
>
> // Wait for the dispatch i.e. response has been returned
> assert(Await.result(service(req)) === req.response)
>
> trans.close()
> intercept[Reader.ReaderDiscarded] { Await.result(req.writer.write(..)) }

RB_ID=276657