[split] ForkJoin scheduler: first draft
A scheduler based on the JSR166y fork/join implementation. This
scheduler will balance tasks (via work stealing) across several
threads while retaining a target level of parallelism.

It also manages blocking, which is threaded through from Await; thus,
with this scheduler, applications can block with only a small penalty,
while avoiding the possibility of deadlock due to the parallelism
invariant ??? progress is guaranteed.

Currently, only future blocking is managed, this may be extended in
the future.

I think we can actually come up with a better work-stealing scheduler
for Finagle, but this is a good first approximation and has shown very
promising results in the adshards.

RB_ID=288821