[split] [finagle-stats] use java.lang.Double in addGauge()
Motivation:
JsonExporter exports stats using jackson's ObjectWriter. This uses
reflection to access the type of Number to print it. For addGauge,
this ends up calling Float.toString, which limits the significant
figures of the output to 8 which is not sufficient for long values.

Modifications:
This change creates gauges of type java.lang.Double instead.

Result:
JsonExporter now calls Double.toString which does not truncate.

RB_ID=312585