[split] Ketama memcache: accept weighted addresses
The Ketama memcache implementation looks specifically for inet
addresses -- it has to, since it has to have something to hash -- but
this precludes any weighted addresses that might be passed in.

We introduce new matchers for weighted address which are liberal:
absent a weight, it sets the weight to 1. We then use this in the
Ketama implementation.

In the future, we may consider propagating these weights all the way
down, since Ketama itself has a concept of weighting.

RB_ID=307519