[split] finagle-core: make StackClient symmetric to StackServer
Problem

StackClient composed more cleanly into a Stack[ServiceFactory]
than a StackServer. However, this made it more opaque and potentially
unsafe as there was no way to verify that a proper codec was
defined.

Solution

This change makes StackClient an abstract class which requires
a Transporter and Dispatcher. This is symmetric to a StackServer.
Additionally, shared params between a StackClient and StackServer
have been factored into appropriate companion objects.

RB_ID=325993