[split] Add Read-Only ZK Support to Finagle-serversets zk2 Resolver
This introduces improvements to the ZK protocol to allow clients to
read from observers even while quorum is unavailable. This means that
serverset clients should be able to get the last-known-good view of
existing endpoints as long as they're able to connect to an observer.
This, in combination with the already existent stabilization features
in serverset2, means that serverset clients now have two layers of
protection from partition events. In addition, we're adding client-
side stats for ZK operations, latency, and session events.

- Update serverset2 to use (package-internal) ZK Facade.
- Allow ZK ClientBuilder to prioritize different implementations.
- Private ZK client-internal for serverset2.
- Adding exceptions added in ZK 3.5.x
- Twitter-server-internal users get client-internal for zk2.

RB_ID=319901