[split] [finagle-core] transport should be considered closed if it is failed
Problem

When failing the channel transport, the channel is closed and its
close promise completed. Closing the channel is asynchronous though
so the trasnport still is considered open until the close operation
completes.

Solution

Check the status of "failed" when checking if the channel is open.

Result

Channels that are failing will not be considered open. Services
that dispatch to failing channels will not be considered available.

RB_ID=336543