[split] finagle-stats: Enable cumulative gauges in MetricsStatsReceiver
Motivation:
MetricsStatsReceiver was lacking support for cumulative gauges.
Indeed there's a few usage of that in finagle, when two code path create
a gauge with the same name, they're summed during the collection.

Modifications:
- MetricsStatsReceiver now inherits StatsReceiverWithCumulativeGauges
- Port Spec tests to scalatest
- Add a new test for cumulative gauges
- Avoid sorting the stats on "non-pretty-printed" collection

Result:
Adding two gauges with the same name results in one gauge, its value is the sum
of the two previous one.

RB_ID=326301