[split] Mux: reflect reality in specification
While implementing Mux in another context and language, I discovered
an inconsistency.

The Mux specification did not accurately reflect the implementation in
the description of message types. The specification is "correct" --
this was a silly bug -- but unfortunately it is too late to correct
without bumping the version number. So we'll wait.

In the meantime, we can at least create a definition which is
consistent, and compensate for corner cases in the implementation.
Once new clients and servers are out, we can remove the aliases
introduced here.

RB_ID=307247