[split] finagle-core: fix race case in DelayedFactory
Problem

DelayedFactory can throw set a Promise twice if completion of the underlying factory races with an interrupt to a waiter.
The race occurs when the interrupt observes and retrieves the queue of waiters, then the underlying factory completes, then the element is removed from the queue and set as an exception.

Solution

I've changed the interrupt handler to synchronize before reading the state. An alternative solution to this specific problem is to empty the queue inside of the underlying completion handler, but that still poses the issue that the interrupt handler could be acting on an out of date state.

Result

The Promise is no longer set twice.

RB_ID=365285