[split] finagle-mux: Rm ClientHangupException in favor of CancelledRequestException
Problem

Mux introduces com.twitter.finagle.mux.ClientHangupException for use
when raising service Futures upon failure of a downstream request. In
the non-Mux ServerDispatcher, the equivalent exception type is
com.twitter.finagle.CancelledRequestException. Since every success rate
chart and alert in the company is written in terms of this latter
exception type, it makes sense to reuse it in Mux in order to avoid
having to rewrite them all.

Solution

Remove ClientHangupException and replace its usages with
CancelledRequestExceptions. This required adding a `cause: Throwable`
constructor argument to CancelledRequestException. The no-arg
constructor is also maintained for backwards compatibility.

RB_ID=366059