Initial draft of HTTP streaming documentation
Problem

It's too difficult to learn HTTP streaming by reading the source, and
the test cases in EndToEndTests.scala are too alien looking to model
real world problems on.

Solution

Documentation and examples of usage.

Result

More people know about HTTP streaming and will be comfortable to try it
out.