[split] Removing specs from the CookieMapSpec test.
Problem

In order to upgrade to 2.11 (and as part of a larger twitter initiative)
specs has to be removed. This is one small step in doing so.

Solution

Remove specs, replace with the standard FunSuite / assert way of
testing.

Result.

Hopefully nothing will change. Except for removing specs.

Signed-off-by: Moses Nakamura <mnakamura@twitter.com>

RB_ID=400183
TBR=true