[split] [finagle-core] add a generic DtabStatsFilter
Problem:

We need a way to track Dtab usage in finagle-http clients; however this would be generically useful in servers and in thrift applications as well.

Solution:

Provide a generic DtabStatsFilter that tracks the size (and count) of Dtab.local for requests with non-empty dtabs.

Result:

There is a mechanism to track stats of Dtab.local usage.

RB_ID=413631