[split] finagle-core: Forces finagle-core to use ipv4 network stack
Problem

Running tests with maven locally, we use ipv6 by default, but with
pants, we use ipv4.  This means that there are tests which work in
pants, but fail in maven.

Solution

Unification is the first step.  Right now, we're moving maven over
to ipv4, since it's the path of least resistance.  Later, it would
be nice to be unified on ipv6, but that seems like it will take a
while.

Result

finagle-core tests run with ipv4, and pass under both maven and
pants.

RB_ID=420433