[split] util: Enables cross-publishing for 2.11
Problem

Although the specs tests were removed, we weren't publishing against
2.11.  Notably, util-eval will be killed for 2.11, and it's nontrivial
to get sbt to aggregate a project for some crossScalaVersions, and
not for others.  Also, we were depending on a scalacheck feature
that was removed in newer versions of scalacheck.

Solution

Upgraded scalacheck version for 2.9.2 to one that had a replacement
method that was the same in the up to date scalacheck.  Moved
util-eval out from being aggregated, and changed its crossScalaVersion
to only "2.9.2", "2.10.4".  It turned out that the version of sbt
we were on had a bug where it didn't use incremental compilation
if you used cross building, so I upgraded us to 0.13.5 across the
board.  util-logging was ooming, so I increased the heap size.

Result

We can now publish against 2.11

RB_ID=425979