[split] fix com.twitter.finagle.serverset2.ZkResolverTest by increasing time tolerance for node removal
Problem:

com.twitter.finagle.serverset2.ZkResolverTest fails reliably. It is currently marked as flaky, but it never works. This is because, when a node is removed from zk, it takes over 40 seconds for the change to be observed by a resolver (due to stabilization logic).

Solution:

Increase time tolerances for removal checks. Leave the test to be marked as flaky, since it relies heavily on timing.

Result:

finagle-serverset tests pass without setting SKIP_FLAKY.

RB_ID=440149