[split] finagle-thriftmux: Record protocol as a gauge rather than a counter
Problem

Finagle currently records ThriftMux protocol usage with counters. This
is hard to track cluster-wide because the counters appear on charts as
single blips on service startup.

Solution

We persist a single gauge per server and client which holds a value of
1.0. By using gauges instead of counters, cluster-wide ThriftMux usage
will be easily trackable in the same fashion as are connections. Since
the metric is recorded within a `Stackable.make` function, we can't
val up the gauge and instead must use `provideGauge`.

RB_ID=442059